requirements.txt
altgraph==0.17.5
bottle==0.13.4
cffi==2.0.0
clr-loader==0.2.10
nuitka==4.0
packaging==26.0
pefile==2024.8.26
proxy-tools==0.1.0
pycparser==3.0
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
pythonnet==3.0.5
pywebview==6.1
pywin32-ctypes==0.2.3
setuptools==82.0.0
typing-extensions==4.15.0
zstandard==0.25.0
